issue 2768 refactor api error handling + more informative err 400
