+++ extension/chrome/elements/compose-modules/compose-draft-module.ts
@@ -4,7 +4,7 @@
 
 import { Mime, MimeContent, MimeProccesedMsg } from '../../../js/common/core/mime.js';
 import { AcctStore } from '../../../js/common/platform/store/acct-store.js';
-import { AjaxErr } from '../../../js/common/api/error/api-error-types.js';
+import { AjaxErr } from '../../../js/common/api/error/api-error.js';
 import { ApiErr } from '../../../js/common/api/error/api-error.js';
 import { BrowserMsg } from '../../../js/common/browser/browser-msg.js';
 import { Buf } from '../../../js/common/core/buf.js';
